Edit: moved the two main classes outside
